+++ src/templates.rs
@@ -149,3 +149,94 @@ pub fn mfa_configured_mail(mfa_type: String) -> Result<String, TemplateError> {
 
     Ok(tera.render("mail_mfa_configured", &context)?)
 }
+
+#[cfg(test)]
+mod test {
+    use claims::assert_ok;
+
+    use super::*;
+
+    fn get_welcome_context() -> Context {
+        let mut context = Context::new();
+        context.insert("first_name", "test_first");
+        context.insert("last_name", "test_last");
+        context.insert("username", "username");
+        context.insert("defguard_url", "test_url");
+        context.insert("defguard_version", &VERSION);
+        context.insert("admin_first_name", "test_first_name");
+        context.insert("admin_last_name", "test_last_name");
+        context.insert("admin_email", "test_email");
+        context.insert("admin_phone", "test_phone");
+        context
+    }
+
+    #[test]
+    fn test_base_mail_no_context() {
+        assert_ok!(get_base_tera(None));
+    }
+
+    #[test]
+    fn test_base_mail_external_context() {
+        let external_context: Context = Context::new();
+        assert_ok!(get_base_tera(Some(external_context)));
+    }
+
+    #[test]
+    fn test_test_mail() {
+        assert_ok!(test_mail());
+    }
+
+    #[test]
+    fn test_enrollment_start_mail() {
+        assert_ok!(enrollment_start_mail(
+            Context::new(),
+            Url::parse("http://localhost:8080").unwrap(),
+            "test_token"
+        ));
+    }
+
+    #[test]
+    fn test_enrollment_welcome_mail() {
+        assert_ok!(enrollment_welcome_mail("Hi there! Welcome to DefGuard."));
+    }
+
+    #[test]
+    fn test_desktop_start_mail() {
+        let external_context = get_welcome_context();
+        let url = Url::parse("http://127.0.0.1:8080").unwrap();
+        let token = "TestToken";
+        assert_ok!(desktop_start_mail(external_context, url, token));
+    }
+
+    #[test]
+    fn test_new_device_added_mail() {
+        let template_locations: Vec<TemplateLocation> = vec![
+            TemplateLocation {
+                name: "Test 01".into(),
+                assigned_ip: "10.0.0.10".into(),
+            },
+            TemplateLocation {
+                name: "Test 02".into(),
+                assigned_ip: "10.0.0.10".into(),
+            },
+        ];
+        assert_ok!(new_device_added_mail(
+            "Test device",
+            "TestKey",
+            &template_locations
+        ));
+    }
+
+    #[test]
+    fn test_enrollment_admin_notification() {
+        let test_user: User = User::new(
+            "test".into(),
+            "1234".into(),
+            "test_last".into(),
+            "test_first".into(),
+            "test@example.com".into(),
+            Some("99999".into()),
+        );
+        assert_ok!(enrollment_admin_notification(&test_user, &test_user));
+    }
+}
